dom/serviceworkers/test/test_eval_allowed.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1160458 - CSP activated by default in Service Workers</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">
  function register() {
    return navigator.serviceWorker.register("eval_worker.js");
  }

  function runTest() {
    try {
      // eslint-disable-next-line no-eval
      eval("1");
      ok(false, "should throw");
    }
    catch (ex) {
      ok(true, "did throw");
    }
    register()
      .then(function(swr) {
        ok(true, "eval restriction didn't get inherited");
        swr.unregister()
          .then(function() {
            SimpleTest.finish();
          });
      }).catch(function(e) {
        ok(false, "Some test failed with error " + e);
        SimpleTest.finish();
      });
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.exemptFromPerDomainMax", true],
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true],
  ]}, runTest);
</script>
</pre>
</body>
</html>
